+++ include/arg_callback.h
@@ -5,7 +5,7 @@
 ** Login  <leroy_v@epitech.eu>
 **
 ** Started on  Sun Apr 05 16:29:01 2015 vincent leroy
-** Last update Tue Apr 07 17:15:43 2015 vincent leroy
+** Last update Thu Apr 30 15:34:24 2015 vincent leroy
 */
 
 #ifndef ARG_CALLBACK_H_
@@ -19,5 +19,6 @@ bool set_thread_number(const char *option, const char *arg, void *userdata);
 bool set_memlimit(const char *option, const char *arg, void *userdata);
 bool set_compress(const char *option, const char *arg, void *userdata);
 bool set_decompress(const char *option, const char *arg, void *userdata);
+bool set_test(const char *option, const char *arg, void *userdata);
 
 #endif /* !ARG_CALLBACK_H_ */