+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * Threads in the threadpool are canceled then deleted
 * read_raw_data function read mem_size bytes
+* Add the 'test' option to check the archive integrity
 
 1.0.0 / 26-04-2015
 ------------------